Create searching_Keyword
adding a file to search a keyword with a button at the top right of the webpage. everytime one clicks on the button the keyword "car" will be searched on the webpage and it will be highlighted for the user to see where in the webpage the specific keyword exists or not.